requirements.txt
transformers==4.28.1
torch>=1.13.1
moviepy==1.0.3
fpdf==1.7.2
Pillow==9.5.0
yt-dlp
ffmpeg-python==0.2.0
numpy==1.23.5
